pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.1"
description = "finds the difference between files"
authors = ["weii3hai <weii3hai@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.8"
flake8 = "^5.0.4"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
PyYAML = "^6.0"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
